requirements.txt
streamlit 
google-generativeai
PyPDF2
python-dotenv
huggingface-hub==0.20.1
langchain==0.0.350
langchain-community==0.0.3
replicate
opencv-python
stqdm
pdf2image
moviepy
ffmpeg
openai
opencv-contrib-python